+++ serlib/ser_environ.ml
@@ -8,13 +8,14 @@
 
 (************************************************************************)
 (* Coq serialization API/Plugin                                         *)
-(* Copyright 2016-2018 MINES ParisTech                                  *)
+(* Copyright 2016-2019 MINES ParisTech                                  *)
 (************************************************************************)
 (* Status: Experimental                                                 *)
 (************************************************************************)
 
-open Sexplib.Std
+open Sexplib.Conv
 
+module CEphemeron = Ser_cEphemeron
 module Range  = Ser_range
 module Names  = Ser_names
 module Constr = Ser_constr
@@ -38,10 +39,45 @@ type named_context_val =
   [%import: Environ.named_context_val]
   [@@deriving sexp_of]
 
-type globals =
-  [%import: Environ.globals]
+type link_info =
+  [%import: Environ.link_info]
+  [@@deriving sexp]
 
-let sexp_of_globals = Serlib_base.sexp_of_opaque ~typ:"Environ.globals"
+(* For 4.06 *)
+module Pervasives = struct
+  type nonrec 'a ref = 'a ref
+    [@@deriving sexp]
+end
+
+(* For 4.07 *)
+module Stdlib = struct
+  type nonrec 'a ref = 'a ref
+    [@@deriving sexp]
+end
+
+type key = 
+  [%import: Environ.key]
+  [@@deriving sexp]
+
+type constant_key = 
+  [%import: Environ.constant_key]
+  [@@deriving sexp]
+
+type mind_key =   
+  [%import: Environ.mind_key]
+  [@@deriving sexp]
+
+type _globals = {
+  env_constants : constant_key Names.Cmap_env.t;
+  env_inductives : mind_key Names.Mindmap_env.t;
+  env_modules : Declarations.module_body Names.MPmap.t;
+  env_modtypes : Declarations.module_type_body Names.MPmap.t;
+} [@@deriving sexp]
+
+type globals = Environ.globals
+
+let sexp_of_globals g = sexp_of__globals Obj.(magic g)
+let _globals_of_sexp g = Obj.magic (_globals_of_sexp g)
 
 type env =
   [%import: Environ.env]